fix: 重複していた関数を修正
