+++ frontend/src/features/auth/hooks/utils.ts
@@ -3,8 +3,15 @@
  * @param buffer - 変換対象のバイナリデータ
  * @returns Base64エンコードされた文字列
  */
-export const arrayBufferToBase64 = (buffer: ArrayBuffer): string => {
-  return btoa(String.fromCodePoint(...new Uint8Array(buffer)))
+export const arrayBufferToBase64 = (
+  buffer: ArrayBuffer | ArrayBufferLike,
+): string => {
+  const uint8Array =
+    buffer instanceof Uint8Array ? buffer : new Uint8Array(buffer)
+  return btoa(String.fromCodePoint(...uint8Array))
+    .replaceAll("+", "-")
+    .replaceAll("/", "_")
+    .replaceAll("=", "")
 }
 
 /**
@@ -13,10 +20,20 @@ export const arrayBufferToBase64 = (buffer: ArrayBuffer): string => {
  * @returns デコードされたバイナリデータ
  */
 export const base64ToArrayBuffer = (base64: string): ArrayBuffer => {
-  const binaryString = atob(base64)
-  const bytes = new Uint8Array(binaryString.length)
-  for (let i = 0; i < binaryString.length; i++) {
-    bytes[i] = String.prototype.codePointAt.call(binaryString, i) as number
+  try {
+    const standardBase64 = base64
+      .replaceAll("-", "+")
+      .replaceAll("_", "/")
+      .padEnd(base64.length + ((4 - (base64.length % 4)) % 4), "=")
+
+    const binaryString = atob(standardBase64)
+    const bytes = new Uint8Array(binaryString.length)
+    for (let i = 0; i < binaryString.length; i++) {
+      bytes[i] = String.prototype.codePointAt.call(binaryString, i) as number
+    }
+    return bytes.buffer as ArrayBuffer
+  } catch (error) {
+    console.error("Base64 decoding error:", error)
+    throw new Error("Base64文字列のデコードに失敗しました")
   }
-  return bytes.buffer as ArrayBuffer
 }
